Add is_renting to Client Model
Added property to get track if a client has a rented car.
Client can only has one car at time.
The app is not going to check calendar combinations and superpositions.

The transactions are recorded day to day, meaning that a car has
to return to place before it can be assign to a new client.